+++ internal/provider/resource_domain.go
@@ -12,12 +12,11 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/vmware/terraform-provider-vcf/internal/cluster"
 	"github.com/vmware/terraform-provider-vcf/internal/constants"
+	"github.com/vmware/terraform-provider-vcf/internal/domain"
 	"github.com/vmware/terraform-provider-vcf/internal/network"
 	"github.com/vmware/terraform-provider-vcf/internal/resource_utils"
 	validationUtils "github.com/vmware/terraform-provider-vcf/internal/validation"
 	"github.com/vmware/terraform-provider-vcf/internal/vcenter"
-	"github.com/vmware/vcf-sdk-go/client"
-	"github.com/vmware/vcf-sdk-go/client/clusters"
 	"github.com/vmware/vcf-sdk-go/client/domains"
 	"github.com/vmware/vcf-sdk-go/models"
 	"reflect"
@@ -30,7 +29,16 @@ func ResourceDomain() *schema.Resource {
 		ReadContext:   resourceDomainRead,
 		UpdateContext: resourceDomainUpdate,
 		DeleteContext: resourceDomainDelete,
-		// TODO implement wld import, but fail import of Management domain
+		Importer: &schema.ResourceImporter{
+			StateContext: func(ctx context.Context, data *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+				vcfClient := meta.(*SddcManagerClient)
+				apiClient := vcfClient.ApiClient
+				domainId := data.Id()
+				// NOTE: Management domain cannot be imported, to not allow users to accidentally delete it,
+				// but it can be used as datasource
+				return domain.ImportDomain(ctx, data, apiClient, domainId, false)
+			},
+		},
 		Timeouts: &schema.ResourceTimeout{
 			Create: schema.DefaultTimeout(4 * time.Hour),
 			Read:   schema.DefaultTimeout(20 * time.Minute),
@@ -50,7 +58,7 @@ func ResourceDomain() *schema.Resource {
 				ValidateFunc: validation.StringLenBetween(3, 20),
 				Description:  "Organization name of the workload domain",
 			},
-			"vcenter": {
+			"vcenter_configuration": {
 				Type:        schema.TypeList,
 				Required:    true,
 				Description: "Specification describing vCenter Server instance settings",
@@ -105,7 +113,7 @@ func resourceDomainCreate(ctx context.Context, data *schema.ResourceData, meta i
 	vcfClient := meta.(*SddcManagerClient)
 	apiClient := vcfClient.ApiClient
 
-	domainCreationSpec, err := createDomainCreationSpec(data)
+	domainCreationSpec, err := domain.CreateDomainCreationSpec(data)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -148,35 +156,19 @@ func resourceDomainRead(ctx context.Context, data *schema.ResourceData, meta int
 	vcfClient := meta.(*SddcManagerClient)
 	apiClient := vcfClient.ApiClient
 
-	getDomainParams := domains.NewGetDomainParamsWithContext(ctx).
-		WithTimeout(constants.DefaultVcfApiCallTimeout)
-	getDomainParams.ID = data.Id()
-	domainResult, err := apiClient.Domains.GetDomain(getDomainParams)
+	domainObj, err := domain.SetBasicDomainAttributes(ctx, data.Id(), data, apiClient)
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	domain := domainResult.Payload
 
-	_ = data.Set("name", domain.Name)
-	_ = data.Set("status", domain.Status)
-	_ = data.Set("type", domain.Type)
-	_ = data.Set("sso_id", domain.SSOID)
-	_ = data.Set("sso_name", domain.SSOName)
-	_ = data.Set("is_management_sso_domain", domain.IsManagementSSODomain)
-	if len(domain.VCENTERS) < 1 {
-		return diag.FromErr(fmt.Errorf("no vCenter Server instance found for domain %q", data.Id()))
-	}
-
-	vcenterConfigRaw := data.Get("vcenter").([]interface{})
-	vcenterConfig := vcenterConfigRaw[0].(map[string]interface{})
-	vcenterConfig["id"] = domain.VCENTERS[0].ID
-	vcenterConfig["fqdn"] = domain.VCENTERS[0].Fqdn
-	_ = data.Set("vcenter", vcenterConfigRaw)
-
-	err = readAndSetClustersDataToDomainResource(domain.Clusters, ctx, data, apiClient)
+	err = domain.ReadAndSetClustersDataToDomainResource(domainObj.Clusters, ctx, data, apiClient)
 	if err != nil {
 		return diag.FromErr(err)
 	}
+	nsxtClusterConfigRaw := data.Get("nsx_configuration").([]interface{})
+	nsxtClusterConfig := nsxtClusterConfigRaw[0].(map[string]interface{})
+	nsxtClusterConfig["id"] = domainObj.NSXTCluster.ID
+	_ = data.Set("nsx_configuration", nsxtClusterConfigRaw)
 
 	return nil
 }
@@ -187,7 +179,7 @@ func resourceDomainUpdate(ctx context.Context, data *schema.ResourceData, meta i
 
 	// Domain Update API supports only changes to domain name and Cluster Import
 	if data.HasChange("name") {
-		domainUpdateSpec := createDomainUpdateSpec(data, false)
+		domainUpdateSpec := domain.CreateDomainUpdateSpec(data, false)
 		domainUpdateParams := domains.NewUpdateDomainParamsWithContext(ctx).
 			WithTimeout(constants.DefaultVcfApiCallTimeout)
 		domainUpdateParams.DomainUpdateSpec = domainUpdateSpec
@@ -293,7 +285,7 @@ func resourceDomainDelete(ctx context.Context, data *schema.ResourceData, meta i
 	vcfClient := meta.(*SddcManagerClient)
 	apiClient := vcfClient.ApiClient
 
-	markForDeleteUpdateSpec := createDomainUpdateSpec(data, true)
+	markForDeleteUpdateSpec := domain.CreateDomainUpdateSpec(data, true)
 	domainUpdateParams := domains.NewUpdateDomainParamsWithContext(ctx).
 		WithTimeout(constants.DefaultVcfApiCallTimeout)
 	domainUpdateParams.DomainUpdateSpec = markForDeleteUpdateSpec
@@ -329,136 +321,4 @@ func resourceDomainDelete(ctx context.Context, data *schema.ResourceData, meta i
 	}
 
 	return nil
-}
-
-func createDomainCreationSpec(data *schema.ResourceData) (*models.DomainCreationSpec, error) {
-	result := new(models.DomainCreationSpec)
-	domainName := data.Get("name").(string)
-	result.DomainName = &domainName
-
-	if orgName, ok := data.GetOk("org_name"); ok {
-		result.OrgName = orgName.(string)
-	}
-
-	vcenterSpec, err := generateVcenterSpecFromResourceData(data)
-	if err == nil {
-		result.VcenterSpec = vcenterSpec
-	} else {
-		return nil, err
-	}
-
-	nsxSpec, err := generateNsxSpecFromResourceData(data)
-	if err == nil {
-		result.NsxTSpec = nsxSpec
-	} else {
-		return nil, err
-	}
-
-	computeSpec, err := generateComputeSpecFromResourceData(data)
-	if err == nil {
-		result.ComputeSpec = computeSpec
-	} else {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-func readAndSetClustersDataToDomainResource(domainClusterRefs []*models.ClusterReference,
-	ctx context.Context, data *schema.ResourceData, apiClient *client.VcfClient) error {
-	clusterIdsInTheCurrentDomain := make(map[string]bool, len(domainClusterRefs))
-	for _, clusterReference := range domainClusterRefs {
-		clusterIdsInTheCurrentDomain[*clusterReference.ID] = true
-	}
-
-	getClustersParams := clusters.NewGetClustersParamsWithContext(ctx).
-		WithTimeout(constants.DefaultVcfApiCallTimeout)
-
-	clustersResult, err := apiClient.Clusters.GetClusters(getClustersParams)
-	if err != nil {
-		return err
-	}
-	domainClusterData := data.Get("cluster")
-	domainClusterDataList := domainClusterData.([]interface{})
-	allClusters := clustersResult.Payload.Elements
-	for _, domainClusterRaw := range domainClusterDataList {
-		domainCluster := domainClusterRaw.(map[string]interface{})
-		for _, clusterObj := range allClusters {
-			_, ok := clusterIdsInTheCurrentDomain[clusterObj.ID]
-			// go over clusters that are in the domain, skip the rest
-			if !ok {
-				continue
-			}
-			if domainCluster["name"] == clusterObj.Name {
-				domainCluster["id"] = clusterObj.ID
-				domainCluster["primary_datastore_name"] = clusterObj.PrimaryDatastoreName
-				domainCluster["primary_datastore_type"] = clusterObj.PrimaryDatastoreType
-				domainCluster["is_default"] = clusterObj.IsDefault
-				domainCluster["is_stretched"] = clusterObj.IsStretched
-			}
-		}
-	}
-	_ = data.Set("cluster", domainClusterData)
-
-	return nil
-}
-
-func createDomainUpdateSpec(data *schema.ResourceData, markForDeletion bool) *models.DomainUpdateSpec {
-	result := new(models.DomainUpdateSpec)
-	if markForDeletion {
-		result.MarkForDeletion = true
-		return result
-	}
-	if data.HasChange("name") {
-		result.Name = data.Get("name").(string)
-	}
-
-	// TODO implement support for IPPoolSpecs in NsxTSpec
-	// by placing the added cluster spec in the DomainUpdateSpec
-	//nsxtSpec, err := generateNsxSpecFromResourceData(data)
-	//if err == nil {
-	//	result.NsxTSpec = nsxtSpec
-	//} else {
-	//	return nil, err
-	//}
-
-	return result
-}
-
-func generateNsxSpecFromResourceData(data *schema.ResourceData) (*models.NsxTSpec, error) {
-	if nsxConfigRaw, ok := data.GetOk("nsx_configuration"); ok && len(nsxConfigRaw.([]interface{})) > 0 {
-		nsxConfigList := nsxConfigRaw.([]interface{})
-		nsxConfigListEntry := nsxConfigList[0].(map[string]interface{})
-		nsxSpec, err := network.TryConvertToNsxSpec(nsxConfigListEntry)
-		return nsxSpec, err
-	}
-	return nil, nil
-}
-
-func generateVcenterSpecFromResourceData(data *schema.ResourceData) (*models.VcenterSpec, error) {
-	if vcenterConfigRaw, ok := data.GetOk("vcenter"); ok && len(vcenterConfigRaw.([]interface{})) > 0 {
-		vcenterConfigList := vcenterConfigRaw.([]interface{})
-		vcenterConfigListEntry := vcenterConfigList[0].(map[string]interface{})
-		vcenterSpec, err := vcenter.TryConvertToVcenterSpec(vcenterConfigListEntry)
-		return vcenterSpec, err
-	}
-	return nil, nil
-}
-
-func generateComputeSpecFromResourceData(data *schema.ResourceData) (*models.ComputeSpec, error) {
-	if clusterConfigRaw, ok := data.GetOk("cluster"); ok && !validationUtils.IsEmpty(clusterConfigRaw) {
-		clusterConfigList := clusterConfigRaw.([]interface{})
-		result := new(models.ComputeSpec)
-		var clusterSpecs []*models.ClusterSpec
-		for _, clusterConfigListEntry := range clusterConfigList {
-			clusterSpec, err := cluster.TryConvertToClusterSpec(clusterConfigListEntry.(map[string]interface{}))
-			if err != nil {
-				return nil, err
-			}
-			clusterSpecs = append(clusterSpecs, clusterSpec)
-		}
-		result.ClusterSpecs = clusterSpecs
-		return result, nil
-	}
-	return nil, fmt.Errorf("no cluster configuration")
 }